+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * added `BinaryParser.get_remaining_bytes`
 * added `BinaryParser.get_remaining_bytes_count`
+* added `BinaryParser.get_parser`